Feat: 추가 ViewerBook [S08P22A101-498]

+++ ItsGrey/data/book/src/main/java/com/tntt/book/datasource/RemoteBookDataSource.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.Flow
 interface RemoteBookDataSource{
     suspend fun createBookDto(userId: String, bookDto: BookDto): Flow<String>
     suspend fun getBookDto(bookId: String): Flow<BookDto>
-    suspend fun getBookDtoList(userId: String, sortType: SortType, startIndex: Long, bookType: BookType): Flow<List<BookDto>>
+    suspend fun getBookDtoList(userId: String, sortType: SortType, startIndex: Int, bookType: BookType): Flow<List<BookDto>>
     suspend fun updateBookDto(bookDto: BookDto): Flow<Boolean>
     suspend fun deleteBookDto(bookIdList: List<String>): Flow<Boolean>
 }
